localcosmos_server/requirements.txt
django==3.1.*
djangorestframework==3.11.1
django-imagekit==4.0.2
django-road
content-licencing
anycluster
rules==2.2
google-cloud-vision
django-el-pagination==3.3.0
django-octicons==1.0.2
django-countries==6.1.3
django-cors-headers==3.5.0
Pillow
matplotlib
